@@ -1,31 +1,14 @@
-import { Inject, Injectable } from '@angular/core';
-import {
-  HttpRequest,
-  HttpHandler,
-  HttpEvent,
-  HttpInterceptor
-} from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { HttpInterceptorFn } from '@angular/common/http';
+import { inject } from '@angular/core';
 import { OKTA_AUTH } from '@okta/okta-angular';
-import { OktaAuth } from '@okta/okta-auth-js';
 
-@Injectable()
-export class AuthInterceptor implements HttpInterceptor {
-
-  constructor(@Inject(OKTA_AUTH) private _oktaAuth: OktaAuth) {}
-
-  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
-    return next.handle(this.addAuthHeaderToAllowedOrigins(request));
+export const authInterceptor: HttpInterceptorFn = (req, next, oktaAuth = inject(OKTA_AUTH)) => {
+  let request = req;
+  const allowedOrigins = ['http://localhost'];
+  const accessToken = oktaAuth.getAccessToken();
+  if(accessToken && !!allowedOrigins.find(origin => req.url.includes(origin))) {
+    request = req.clone({ setHeaders: { 'Authorization': `Bearer ${accessToken}` } });
   }
 
-  private addAuthHeaderToAllowedOrigins(request: HttpRequest<unknown>): HttpRequest<unknown> {
-    let req = request;
-    const allowedOrigins = ['http://localhost'];
-    if(!!allowedOrigins.find(origin => request.url.includes(origin))) {
-      const authToken = this._oktaAuth.getAccessToken();
-      req = request.clone({ setHeaders: { 'Authorization': `Bearer ${authToken}` } });
-    }
-
-    return req;
-  }
-}
+  return next(request);
+};
