@@ -8,7 +8,7 @@ describe('ProtectedComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ProtectedComponent ]
+      imports: [ ProtectedComponent ]
     })
     .compileComponents();
   });
